rename plan/src/plan.rs to plan/src/model.rs
